RakuAST: allow for =formula, as per rakudoc 2.0
The reason this didn't work is because =formula shares the prefix
with =for.  Fixed for now by adding a specific token for =formula.

Turns out the legacy grammar **also** has an issue with any
abbreviated rakudoc items that share the prefix with a directive,
such as =beginna and =enda.  I wonder why nobody found this to be
a problem before in the past 10+ years: I guess a general fix is
just not that easy, *or* would slow down parsing in general too
much.