Make sure Arrays also get hammered
The Array class had a shortcut to .flat, because all of the elements
in an Array are always containerized, so calling .flat on it wouldn't
change anything anyhow.  Until we added the :$hammer named argument.

So if that is specified with a True value, call the Iterable's version
of .flat instead