RakuAST: introducing RakuAST::Deparse::Highlight
This exports a single sub "highlight" that provides Raku-based syntax
highlighting based on first parsing the given code to RakuAST objects
using the Raku grammar with a dedicated set of actions that *also*
process any inline comments.

The resulting RakuAST nodes are then deparsed with a subclass of the
basic RakuAST deparsing class that takes the additional comments
information and provides additional deparsing logic for the inline
comments.

The "highlight" sub takes 2 or more positional arguments: the first is
the source code for which to provide syntax highlighting.

The other arguments indicate the type of highlighting wanted.  They
can either be a string (e.g. "HTML") which will be converted to the
module name "RakuAST::Deparse::Highlight::HTML" and required.  Or
it can be a role that should be mixed in into the Deparse logic.

Since the additional logic is pretty small, this could also become
just part of core.  On the other hand, syntax highlighting is only
needed in very specific cases, so putting it into an ad-hoc loadable
module feels as the right thing to do.

This is alpha quality code: using this in rendering of the Raku
documentation examples, will no doubt expose problems in this logic
and/or the general deparsing logic.