Partially fix mixed bin/no bin reads in IO::Socket
When mixing binary and non-binary accesses, the data is put into an
encoder for decoded accesses. It's possible that not all data that was
added to the decoder is read. Subsequent binary reads would miss the
data already added to the decoder. Fix this by first emptying the decoder
in binary accesses should one be set.

One edge case remains though:
This only works within the limitations of the pipeline of the decoder
itself. In the situation of a UTF-8 char that might still receive
combiners, the decoder might keep bytes in it's pipeline, resulting in
a similar situation than before: the binary read will not see those
bytes stuck in the pipeline and read the bytes after that. Only the
next decoded read will see those bytes.