Allow subroutines as values in enums (part 1)
The problem described in https://github.com/Raku/old-issue-tracker/issues/5818
was caused by trying to stringify the value of the enum to be used
as a key in the reverse lookup of enums.  Now checks if the .WHICH
method can be called on the value, and then uses that as the key.

Note that this does *NOT* fix the issue completely: the enum value is
exported with the given name (without any '&' prefixed).  Then
attempting to call the enum confuses the parser.  So maybe the
enum should be exported with a '&' prefixed, or some other fix
would be needed.