RakuAST: introduce ::Var::Attribute::Public
The syntax "$.foo" is generally thought of as accessing that attribute
in a class, but it is really just "self.foo.item" and was as such
codegenned in the Actions.

This has the disadvantage it not being able to be deparsed easily.
And it also does not provide an easy entry point for "manual" creation
of a RakuAST object representing "$.foo".

The RakuAST::Var::Attribute::Public class is a very thin wrapper around
the "self.foo.item" logic, but allows deparsing and it provides an easy
entry point.

Adjust the actions accordingly, add .raku and deparsing logic and a
simple test.  nine++ for helping me when I got stuck on a simple
oversight