RakuAST: move Text/HTML highlighting into file
These modules are so small, there is really not a lot of point
keeping them in separate files.

Also introduce a DEBUG mode of highlighting, currently a no-op.

Default to Text highlighting if someone is watching, otherwise to
DEBUG mode if output is being redirected (as color codes in a file
don't work too well)

This also saves about .5 second on core module installation