RakuAST: slightly optimize .deparse-without-highlighting
If the invocant is already the base class, then we don't need to
mixin anything.

BTW, this is one case where I would like to have syntax for limiting
the invocant to the base class and disallow any subclasses.  Yes, one
could do:

    class A { method a(\SELF where *.WHAT =:= A:) { } }
    class B is A { }
    B.a;
    # Constraint type check failed in binding to parameter 'SELF';
    # expected anonymous constraint to be met but got B (B)

but that would be very verbose and not very dispatch friendly (whereas
support in dispatch would not be affecting execution speed)