Add :by to .min/.max/.minmax
The sub versions of min/max/minmax take a :by named argument to indicate
the comperator.  However, the method versions take an optional positional
argument and ignore any :by specification (because of the implicit *%_
in the method signature.

This can cause silent breakage when refactoring code from subs to methods
as spotted by vendethiel++
  https://irclogs.raku.org/raku/2024-07-18.html#14:27

This adds 3 candidates taking the :by named argument, so that this
type of breakage will not happen when refactoring code.