RakuAST: make attribute initializers compile in correct context
Attribute initializers are transparently wrapped in a method that will
be called by the object's constructor. Previously this method's AST node
was created after we parsed the initializer. However since the method
provides `self` as an implicit term if the initializer tried to use
`self`, it would fail to resolve the symbol.

Fix by splitting off creation of the Method AST node and doing this at
the variable declaration's parse time, so we can enter the method's
scope before parsing the initializer.