Ensure that sunk failed lookups in native arrays always explode
dies-ok { @arr[2] }; with @arr being a shaped string array with only 1 slot did
not always explode because the raku-sink dispatcher only really called
the sink method on the value if it is not Mu's. Without trying to
dereference the StrPosRef in any way the exception would never got
thrown. Fix by special casing NativeRefHOW in raku-sink to make that
spec test less reliant on precise code gen.