RakuAST: have role bodies be generated directly as such
Until now role bodies were first parsed as blocks and a RakuAST::Block
node generated for them. Then we created an appropriate sub and moved
over the block's contents - a messy affair and violation of the object's
privacy.

Turns out we already have all the information at hand at the time we
need it to instead create a Sub directly. It's just a bit of an
unorthodox process as we cannot just parse a sub (syntactically we only
get a block). Having the sub in place right away simplifies the timing
and gets parsing and synthetic ASTs closer together.

To get the timing in the synthetic case fully under control, we use a
new subclass of RakuAST::Sub called RakuAST::RoleBody.