RakuAST: support use fatal
This replaces the previous attempt to support the fatal pragma.
Fatalization is something that needs to happen on the caller side, not
in the callee. An active fatal pragma in a scope means that any
expressions yielding a Failure should throw an exception right there.
Callees returning those Failures may be defined in completely different
scopes without fatal pragma.

This commit uses most of the implementation from the old compiler
frontend, i.e. it's QAST post-processing. This keeps implementation of
the fatal pragma reasonably contained at the expense of optimizability
(though the chance for static optimizations being possible is slim
anyway).

We deviate from the old implementation in that we use the new fatalize
multi for detecting Failure objects. Apart from generating smaller QAST
there is hope that with new-dispatch the multi call can actually become
cheaper than the inline condition - at least post-spesh.