RakuAST: fix getting properties of custom operator inside EVAL
When resolving the operator, we find it in the outer scope and since we
don't know whether it's a constant or not, we resolve to a
RakuAST::Declaration::External which is not a CompileTimeValue.
Nevertheless, we can still check whether we got a value and since it's
an operator, assume its usable.

Fixes: sub infix:<na>($a, $b) is assoc<non> { ... }; EVAL "1 na 2 na 3"
(which now fails with an appropriate message)