RakuAST: fix reduce meta op being overzealous with thunking
When a reduce meta operator's arguments are contained in parentheses we
would treat the Circumfix::Parenthesis as a single argument when
thunking. That's now how it's meant to work, so instead in this case
treat the list's elements as the actual arguments.

This alone is not enough however as the parenthesis could contain no
arguments at all or a single non-list expression. So we have to treat
these cases specially.