Have throws-like if the expected compile time exception is part of a group
The compiler will throw an X::Comp::Group exception when there are
multiple compile time problems. With time, the compiler gets better and
will report more problems and will also get better at reporting those
problems at the end of compilation instead of right away.

Thus when a test is looking for a specific compile time exception, a
change in the compiler may lead to an X::Comp::Group exception appearing
where previously there was just some X::Comp exception. This would not
mean that anything got broken, quite the opposite. Thus the test should
not fail.

This affects only compile-time exceptions, i.e. tests that ensure that
some broken code does not compile successfully. While the Test modules
is used quite ubiquitously, testing the compiler's error reporting is
much rarer and expected to mostly happen in the spec test suite. Thus
the impact of this semantical change is deemed quite low. And even for
tests of Raku modules the above reasoning for why this is more correct
still applies.