RakuAST: fix constant initialized with un-parameterized type
When initializing a constant with .=new we need to use the full,
parameterized type instead of just the base type as invocant. Luckily we
already have infrastructure that makes sure we get the base type for a
Definite (i.e. Int instead of Int:D) while retaining parameterization.

Fixes my Array[Numeric] constant foo .=new 1, 2
to produce an Array[Numeric] instead of an Array