RakuAST: fix terms in signature declarations
When declaring a term in a signature declaration and initializing that
with an assignment, the term initially needs to be bound to a container.
However, afer initialization is complete, that container needs to go
away again, so future assignments remain prohibited.

We just decont terms after initialization. The old frontend uses a lot
more complicated logic in p6recont_ro but let's see how far we can get
with just a decont.

Fixes: my (\x1) = 1; x1 = 2; # to fail on the second assignment