RakuAST: remove counter productive invocant type check on Grammar methods
Grammar methods like tokens or rules are odd in that they are not called
with the grammar itself as an invocant but instead with a Match object.
Thus doing the usual type check on the invocant would prevent calls from
succeeding. The old frontend does not do any type check and so we
shouldn't either.

Fixes: grammar D { our token doo { doo }; }; say "doo" ~~ &D::doo