RakuAST: make multi regexes operational
A proto regex must not be marked onlystar as we would then use the
optimized dispatcher for multis which would not find any candidates.
That's because while a proto regex is a proto routine, the multi
candidates will not be actual multis and will retain their full name
including the :sym colonpair.