RakuAST: fix confusion of subs and regexes in regex tokens
A <something> in a regex first looks for a regex in lexical scope and
only if it cannot find one, it generates a method call on the grammar.
We already had most of this in place, but did not check whether the
lexical we found was an actual regex and instead took any routine.

Fixes: token foo { <pair> } # which called &pair from the setting.