Prevent dispatch errors on Regex.Bool
By deconting the result of the match before calling .Bool on that.
This makes:

    my $code = $*PROGRAM.slurp x 10000;
    my atomicint $seen;
    for $code.lines.batch(256) -> @lines {
        $*SCHEDULER.cue: {
           ++⚛$seen if / clean / for @lines;
        }
    }
    sleep 1; # wait for things to unfold
    say "seen: $seen";

merely produce incorrect results for number of matches seen, rather
than crashing with crazy dispatch errors such as:

Invocant of method 'Bool' must be an object instance of type 'Mu', not
a type object of type 'Nil'.

and:

Cannot resolve caller Bool(Match: ); none of these signatures matches:
    (Match:U $:: *%_ --> Bool::False)
    (Match:D $:: *%_)