[6.e] Provide `is revision-gated("6.x")` trait
This trait is primarily intended for allowing multi candidates
to be considered for dispatch only when the caller's compiler
version is above a certain revision.

By applying the trait to the proto at its minimum compatible
revision, authors can then apply higher revision requirements
to individual multi candidates.

For example, this is used to constrain the new behavior of
`splice` to versions 6.e and above.

This allows the following to DWIM:

    > use v6.e.PREVIEW;
    > my @a = 1,2,3;
    > @a.splice(1,1,$[7,8])
    > dd @a
    # [1,[7,8],3]

Addresses R#5500 (#5500).