RakuAST: fix losing outer context when switching language version in EVAL
When EVALing code we must not load a setting, even when we switch
language versions. Otherwise the EVAL's outer scope would no longer be
visible. This of course means that the EVALed code does not use the
setting one would expect from the language version (this is a hint that
we should probably prohibit use <version> in EVAL entirely).

Fixes: sub foo() { }; EVAL Q[use v6.c; foo];