Fix the new revision-gated dispatch logic for the RakuAST compiler
Without this, almost all tests fail when RAKUDO_RAKUAST=1.

A few different approaches were attempted to ensure that
`$?LANGUAGE-REVISION` would be installed in the lexical scope
that is accessed by `nqp::getlexcaller(...)` in `dispatchers.nqp`
when the caller is inside one of the RakuAST implementation classes.

* `bootstrap.c/BOOTSTRAP.nqp`
* `Raku/ast/rakuast-prologue.nqp`

Those didn't work. So even though this fix does feel a bit _too_
simple, it works well at a low cost.