RakuAST: initial stab at SafeRakuDoc
This introduces a RakuDoc::To::SafeRakuDoc module, that will usually
be invoked with the --rakudoc commandline argument:

  $ RAKUDO_RAKUAST=1 raku --rakudoc=SafeRakuDoc filename.raku

will compile the source in the given file to an AST, and then just
deparse the RakuDoc of the file.

Current limitations: does not do declarator docs correctly yet.
This will take some iterations and looking at to arrive at some
sort of agreement as to how declarator docs will be converted
to safe rakudoc.

Note that the RAKUDO_RAKUAST=1 is only needed until RakuAST becomes
the default grammar for parsing code, which in turn enables the
--rakudoc commandline logic.