make sure we don't throw out cached folders
doesn't seem to be a problem immediately, since when
there is a cache hit, the cache job will just say
"there is already a cache for key blahblah"
instead of trying to upload from an empty folder
and erroring the whole pipeline.

But for the case that it goes wrong for some reason,
let's make sure there's sensible data to re-cache there.