Address a small but significant corner case in the optimizer
(Fixes R#5644 [#5644])

> with 42 { say $_ ~~ Mu:U }
\# True
> my $m = Mu:U; say $m ~~ Mu:U }
\# True

The above was what the optmized QAST was returning, prmpting
this patch.

It makes sense, because in almost all cases the LHS of a smart
match will return `True` when smartmatched against `Mu`.

The optimizer wisely took this into consideration. There's only
one problem, however, and that's that the moment when you are
genuinely checking for `Mu:U` is the one moment where the
optimization fails.

This commit utilizes `nqp::ifnull` where possible to reduce the
overhead, but it's also a bit of a shame that that I couldn't think
of a way to accomplish this without adding a conditional statemnet
to every `maybe_typematch` call in the optimizer.