Remove $?FILE and $?LINE where they don't make much sense
The filename and line will get set by World if the exception is
thrown at BEGIN time, "file" is not a named argument the
X::Comp constructor takes ("filename" would be), so this is
not a regression. And the filename and line from the core setting
is not very helpful, since you also get a stack trace in that case.