Make sure we have a Bool to check against with :exists
d66162a4095ea0afbd0 introduced a short-cut that would make {}:exists
always return false if the key was a Junction.  Fix this by making
sure we have a Bool:D to compare against (which should be optimized
away pretty quickly if we already had a Bool:D).

The alternative would have been to either revert these optimizations,
or to create almost identical candidates handling Junctions specifically.
That felt like a not very DRY approach, so instead have a little
larger bytecode footprint for the fast path.