Restore :$source parameter to load(...) multi
The `:$source` parameter was removed from one of the `load(...)` multi methods because it was presumed to be unused. The problem, however, was that it was not being used and not that the parameter existed. This restores the `:$source` parameter and now correctly passes it along to the next `load(...)` multi method (which does use the parameter).