Rince repeat c05f170446 for < <= > >=
As they suffered from the same issue.

This breaks an in my view incorrect test that was based on the
incorrect semantics of Rat < Rat in the case of denominator 0:

   0/0..0/0 ~~ -1/0..1/0

was: False, is now True as the right hand side is now a "real"
range as far as .ACCEPTS semantics is concerned.