Fix LTA error message on num % / ** num
The problem is really that *returning* a Failure from a sub that has
a native return constraint **dies**, rather than passing on the
Failure.

If however a CONTROL message is used to return a Failure, all is
well.  So work around this issue by using `fail` with the Failure
rather than just returning it.

Fixes #2434