Make Iterable:U.flat do the right thing
This also exposed a bug in Iterable.flat in that it was **NOT**
marked as "nodal".  Now that there's only one proto with "is nodal"
the Iterable:D.flat also became nodal.

This in turn exposed a faulty test in S03-metaops/hyper.t, which
will be fixed shortly.

Fixes #2442