[JVM] Flatten without overflow/underflow issues
If I'm not mistaken, the code relies on a $level parameter of -1
being coerced to a big positive number (due to the usage of uint).
This doesn't work for the JVM backend.

Since the commit that introduced the uint parameter,
https://github.com/rakudo/rakudo/commit/74cfe8e373, mentioned some
confusion for the expression JIT on Intel, I didn't dare to change
the code for MoarVM.