Make Iterable.flat an only again
Turning it into a multi broke Dan, and possibly other modules that
Blin didn't catch.  Turning it back into an only again, required
changes in Range and native array as well.

The problem is really caused by the fact that current consumers ofi
the role put their own "only" methods in there.  Which would clash
with the multi-ness in the role.

It feels like we need to revisit this behaviour when compositing a
class, or a role into another role.

Mostly a revert of 0fc7d45f66344c