Make "AAS" .. "ABS" use standard .succ semantics in 6.e
In 6.d, this actually follows the "AAS" ... "ABS" (sequence)
semantics, only producing 2 values.  In 6.e it now follows the
generic .succ semantics (with all of its magic regarding strings)
producing 27 values.

Fixes #2238

On the implementation: I would have preferred to use the
is revision-gated trait for this, but this produces a ambiguous
dispatch error in 6.e.  So now the 6.e variant is simple marked
as "is default".