Fail gracefully when REPL is initialized without TTY
This patch was originally intended to address R#4378 (#4378)[^1],
but that problem is actually unresolvable as long as starting
`raku` outside of a TTY automatically engages the wait-until-EOF
mode (the same behavior as calling `perl` without `-e`).

(That mode was previously available via `--repl-mode=non-interactive`,
which this patch exposes via `--repl-mode=disabled`).

However, this patch is still usable so that we successfully fail
in other scenarios, such as if `repl()` would be called inside of
Comma's hand-rolled non-TTY "REPL".

[^1]: There is a widespread issue in "Bash for Git for Windows"
in that it is not in itself a proper TTY. The proposed solution
for docker, node, and others is to run the command again under
`winpty`.