Address long-standing discrepancy for != and METAOP_NEGATE
This patch uses revision-gate to migrate v6e to a version of
negation that preserves junctions while accomplishing the same
underlying functionality as planting `not` in front of `$expr`.

Previously:

    one(0,1) != 0 # False

Now:

    one(0,1) != 0 # one(Bool::False, Bool::True)

Also fixed in this patch is the ability to add a revision-gated
method with the same signature as an already existing method.
In such cases, the existing method is also revision-gated, but to
the revision that it was introduced (in practice this will usually
be the same revision as the proto is gated to).

Addresses R#5686 (#5686) / R#3748 (#3748) / R#3874 (#3874)

See also discussion at:
    https://github.com/Raku/problem-solving/issues/319