RakuAST: fix handling of markup in =alias
As described in https://github.com/Raku/RakuDoc-GAMMA/issues/53

The fix is a bit of a hack: take the contents of the =alias
specification minus the lemma, wrap that in a =begin fake / =end fake
block and run .AST on it.  Then remove the outer block and process
the generated AST accordingly.

Also adapt deparsing of =alias to allow for more than one element.

Note that $*DOC-ALIASES will bind the potential List of ::Block
objects to the lemma.