Restore old logic in find_best_dispatchee
The refactoring from 70b8b229fa changed the logic to detect a mismatch
between the expected type and the passed in value when doing the type
check for native parameters. But if I'm not mistaken the new version
got some cases wrong:

* In the first chain of conditions a container with a native num passed
  in would bypass the check for the expected type.

* In the second chain of conditions any native str, int, or uint value
  that matches the expected type, would be diagnosed as a mismatch due
  to the last alternative.

This commit restores the old logic in both places.

Fixes https://github.com/rakudo/rakudo/issues/5688.