Revision gating is not working with `&infix:<ne>`
I've tried all sorts of things, including moving all the sub
declarations into `core_epilogue.rakumod` ... with some approaches the
CORE.setting fails due to things happening too early or (as is the
case with the current commented out code) using `ne` just causes
an infinite hang.

This commit is just to demonstrate that the desired outcome is possible,
should we decide to not bother with revision gating `ne`.

Another option is to wait until RakuAST is compiling the setting. At
that point, we should be able to move revision gating into the compilation
phase itself. That ought to make things simpler.