Type second param of proto for Stash::ASSIGN-KEY
As explained in https://github.com/rakudo/rakudo/issues/5703 the
current version leads to a spectest failure on the JVM backend. Running
https://github.com/Raku/roast/blob/3032bcf423/integration/advent2011-day07.t
eventually Stash::ASSIGN-KEY is called with something that derives from
Metamodel::GrammarHOW (and doesn't type match with Any) as the second
argument.

Typing the second parameter as Mu solves this problem -- and it also
matches the types used in the multi variants of Stash::ASSIGN-KEY.