Make (almost) any Failure in a slice throws immediately
Before this commit, there was difference in handling slices in which
the indices are produced by a lazy or a non-lazy iterator.  That is
because with indices produced by a lazy iterator, the first element
that returns False on EXISTS-POS will indicate the end of the list
of indices.

  my @a = 0; @a[2] = 2;
  say @a[(^3).lazy];  # (0)
  say @a[^3];         # (0 (Any) 2)

This is behavior that has long be discussed and agreed upon (as far
as I know).  However, any index that is completely out of range, such
as -1, will ALSO return False on EXISTS-POS and thus end the list of
indices in a lazy iterator.  And this causes a discrepancy in the
handling of:

  say @a[(2...-1).lazy];  # (2)
  say @a[(2...-1)];       # Index out of range. Is: -1, should be in 0..^Inf

This issue is further complicated by the use of adverbs.

This commit applies the following semantics:

- If an adverb does not use the actual value (basically, anything with
  :exists or :!exists without a :delete) there is no change in behavior.
  Checking for existence should allow out of bound values.

- In any other case, if EXISTS-POS returns False for the given index,
  a call to AT-POS is also done to see if it produces a Failure.  If
  it does, than that Failure is immediately thrown.

Fixes #3658.

This breaks the final test in t/spec/S02-types/lists.t:

  say 'foo'[2..*].elems;

This used to produce 0 because 2 as an index returned False, and thus
stopped the lazy 2..* iterator, producing a list with 0 elements.  Now
throws an X::OutOfRange error, which I consider the correct behaviour as

  say 'foo'[2..3].elems;

also throws an X::OutOfRange error (and has been doing that before this
commit).