Hopefully address $*USER / $*GROUP on Windows
Thanks to Will Coleda's guidance in #5066, this will hopefully give
some meaningful value to $*USER and $*GROUP on Windows systems, albeit
*without* a numerical component.

As I don't have access to a Windows machine, this is untested, but
judging from the "whoami" output, this should be correct.

While working on this, it struck me that we had a race condition on
initial access to $*USER / $*GROUP, so I also made sure that the whole
process is now protected by a lock (without actually trying to prevent
double work, as this is only to prevent the PROCESS:: stash from
getting frazzled by simultaneous adding of keys).

Will mark issue as fixed after confirmation from Windows users.