Add VM.ownup to produce backtraces of all threads and exit
This wraps Timo Paulssen's new "all-thread-bt" MoarVM syscall in
a method on the VM class, and attempts to do the same on other
backends.

That syscall generates a complete backtrace of *all* running threads
and then exits.  This can be very useful while debugging infinilooping
code.

The "ownup" module wraps this functionality in a control-c wrapper,
so that you can create a full backtrace on an infinilooping program
by adding "-Mownup" on the command line, and pressing control-c when
it becomes clear a program is infinilooping.

The name "own up" was chosen as a non-religious alias for "confess".