Make .trans(/ foo / => ...) 1.5x as fast
The same for any .trans(Iterable:D) cases.

The slowdown was caused by the Pair:D candidate redirecting to the
slow path (if the key and value were not both Str:D) that had a
*@changes signature by wrapping the Pair in a List.

The fix is to make the old slurpy candidate non-slurpy, and to add a
slurpy candidate as a catch-all case, with the resulting array being
specified as the value.

Found as part of work on #5488