Issue a warning for .trans(abc => ...)
As that is an unexpected named argument, it will be ignored.  And
since there are no other transformation specifications, it would
silently become a noop.

Since the op probably meant .trans("abc" => ...) (aka, transforming
the letters a, b and c to something else), an unexpected result would
be returned for any string containing a, b or c.

With this commit a warning will be issued in the case of unexpected
named arguments:

  % raku e-e 'dd "foobar".trans( a => "x")'
  Unexpected named variable(s) ':a("x"),' specified with .trans, did you
  mean to specify Pair(s)?
    in block <unit> at -e line 1
  "foobar"

Part of work on #5488