Fix remaining issue with polymod
Commit c6872fb036 fixed several modules, but broke one polymod test

    12.polymod(lazy 14,)

which should produce (12,).Seq because of the lazy iterator.

However, the iterator for (lazy 14,) is a List::Todo, which returns
True if not exhausted in that case, but False when it *is* exhausted.
I believe that to be an error, but will not touch that before the
2024.12 release.

Note that 12.polymod(14 xx *) *does* produce the correct result, so
I will change that test accordingly.

Part of #5726