Only call ensure-sc if there is a descriptor
On the JVM backend this avoids a NullPointerException for the
following code:

  Format.new("%5s")

The call to nqp::getattr returns a null object in this case (for both
backends). On MoarVM this doesn't lead to an error. But it still feels
appropriate to avoid the call if there is no container descriptor.

This would be the first step for fixing
https://github.com/rakudo/rakudo/issues/5729.