Move test file that dies on JVM to t/09-moar
This is really just a bandaid to make "make test" usable again. The
test doesn't work on the JVM backend: Even if I replace the usage of
atomicint with a plain int, there is a compile time error:

  Type check failed for return value; expected Mu but got BASE (BASE)

As noted in the leading comment of the test file, the final destination
of the test should be roast, anyway. So hopefully moving it to
t/09-moar is good enough for now.