Add proper support for regexes to multi-needle .split
Before this commit, having a regex as one of the needles in a
multi-needle split would just take all the needles and interpolate
them into a single regex.  This was fine as long as you didn't
specify any :k, :kv or :p adverb as well.  If you did, then the
key value would always be 0.

However, solving #5488 would require that a multi-needle split
*would* produce the correct key values.

This commit effectively does that by making sure each needle is
handled separately.  This required a bit of a re-imagining of the
internal logic.  So apart from correctly producing the key values,
it also uses a faster way to handle regexes, which causes one
difference in behaviour if :v, :kv or :p is also specified: instead
of a proper Match object, it produces (just) the matching string.

This appears to be ok roast-wise, but may turn out to become an
issue for more complicated Regexen used with multi-needle .trans.
Until then, this appears to be a good middle ground between
functionality and performance.