Final step of re-imaginining .trans
As started with 097d668036a440614f9 and following commits.

This commit removes the LSM class and its handling of more complex
.trans specifications, and replaces it with a complete re-imagination
basically based on reverse engineering the behaviour using the
test-suite and documentation, all instigated by #5488 (which is now
fixed).

Note that the LSM class was one of the oldest pieces of code still
remaining in the setting, going back to its initial implementation
in 2011, predating the concept of iterators, the GLR and the existence
if multi-needle .split which serves as the new workhorse for the
more complicated .trans specifications.

Many of the more complicated .trans specifications are now up to
3x as fast.  This commit also makes 2 TODOed tests in roast pass.