Apply revision-gating to chmod/chown/rmdir/unlink
The idea is that the semantics of providing multiple-path
versions of these subs is a mistake.

As currently implemented, this patch leaves the behavior the same
for all previous language versions and adding a deprecation message
in 6.e.

In addition, we take this opportunity to align the syntax of calling
`chmod` to the syntax of `chown`. The aligned candidate is available
as of `6.e`, while the previous signature begins to emit a
deprecation message.