Allow native arrays to be endpoints for Ranges
This is a (hopefully) temporary workaround for #5760.  The problem
is caused by dispatch / coercion not selecting the correct Range.new
candidate, causing it to go to the default candidates, which would
then just plunk the native arrays as endpoints in the Range object.
Which then later causes several types of issues when wanting to
iterate over the Range.

This will at least be enough for the 2025.01 release, and unbreak
testing on e.g. the IRC::Log::Colabti module