Make Hash.clone actually clone itself
Fixes #1238.  The problem was caused by the fact that Hash.clone
would create a fresh copy of its invcant, instead of cloning it.
Creating a fresh copy is ok if the invocant is of type core Hash,
but if anything is mixed in, then the fresh copy will lose any
additional attributes (as #1238 showed.

Fixed by changing nqp::create(self) to nqp::clone(self)