Make sure cloning a Date works for subclasses
As found in the research for #1238.  Also re-worked some of the
internals to make it more performant in the new-disp situation
(as most of that code was optimized for the pre-new-disp situation).

Makes Date creation (except Date.today) between 25% faster (basic
cases) and 3.5x as fast (for subclasses of Date)